feat: add components to render news feed in react front-end
Implemented NewsFeed, NewsFeedItem components in the React
front-end part with logic to render the news feed and a
single news feed item, respectively. The NewsFeed component
is composed of instances of NewsFeedItem component to render
each of the news article in the feed. It also has the logic
to make the api call that fetches the news data.

The NewsFeedItem component is purely presentational without
any managed state, rendering single news article data in
corresponding JSX elements, passsed in as a prop from the
parent component.

The NewsFeed component has few lines of code commented out,
to switch between urls used to make the api call directly
to News API end-point and the Express.js backend (which in
turn makes the call to News API), to demonstrate the unsafe
and safe ways of interacting with third-party api's using an
api key in a React app.

The component styling is handled through CSS modules, with
corresponding JSX element styling in each component written
in separate files/modules.

BREAKING CHANGE: React front-end uses different api url and
code depending on whether the url targets the News API end-
point or the Express.js backend, to demo the consequences of
including api keys in React environment variables