feat: add middleware to express backend for handling news data get request
Implemented middleware in Express.js backend to handle GET
request mapped to url: /news-data, to in turn make the
required API call to News API end-point to fetch the desired
news data, to be passed backed to client as a JSON response.
So, it just sends the data back to the React front-end client
and not the typical HTML.

The Express.js back-end server has been configured to listen in
on port 3001 for client requests, as the port 3000 is already
taken by the front-end React's web server.

Additional npm packages have been installed (dotenv, cors, axios)
in the node-express backend to ensure the news api is picked up
by the Express server from the environment variables (dotenv),
to make the api call to third-party api end-point (axios) and
support cross-origin requests (cors) where React client on port
3000 makes a request to Express backend on port 3001, a different
port or origin.

BREAKING CHANGE: Express back-end server handles the delegated task
of sending a request to third party News API end-point, processing
the response and in turn catapults the fetched news data as a JSON
response to the React client